+++ Backend/Backend/settings.py
@@ -41,7 +41,7 @@ INSTALLED_APPS = [
     # apps
     'users',
     'admins',
-    'items',
+
     
 
 ]
